doc: introduce maintainers list
This is a document which lists Tarantool's core subsystems with
their maintainers and reviewers.

Reviewers lists should be used to select people to look at any
patch touching the subsystem. The maintainer lists must be used to
find people responsible for the subsystem as a whole. They can
help with design of major changes, can approve and reject them at
any stage.