luajit: proxy -j and -b flags
There are two flags in the LuaJIT useful for debugging purposes: `-j`
and `-b`. However, if you want to check the same Lua code from the
Tarantool, you will need to make some adjustments in the script itself,
as those flags are not present in the Tarantool's CLI.

This patch introduces those flags to the Tarantool, so debugging is
much more convenient now. Flags are working the same as they do in
LuaJIT.

Closes #5541