log: module name in log messages
Now you can see in the logs the name of the module from which the logging
function was called. The name of the module is determined automatically
during the execution of `require('log')` in the module source code. You can
also set the module name yourself using the function `log.new(module_name)`.

In area of #3211