ci: bump actions/checkout from v1 to v2.3.4
Regarding debug_aarch64: we had old git on odroid machines (see
PR #6180), but now we use graviton ones and, it seems, it becomes
irrelevant.

Regarding memtx_allocator_based_on_malloc: it seems, checkout@v1 here is
due to usage of Debian Stretch (with git 2.11) in docker, as in older
release workflow. Updated it in the spirit of PR #5949.

Left actions/checkout@v1 in containers jobs (coverity and jepsen*): it
should be revisited separately.

Left actions/checkout@v1 in perf_* jobs: they run on special machines
and I don't want to dig inside this part of the infrastructure ATM.