lib: use memcpy for unaligned load/store
Now there's a lot of load_*/store_* functions that are designed
for unaligned access to values in data stream. Unfortunately
they are written in a way that makes new compilers to warn about
outside bounds.

Rewrite it in the most safe, cross-platform and efficient way -
using memcpy. Note that memcpy with compile-time-defined size is
recognized by the magority of compilers and they make the most
efficient operation.

Closes #6703