test: fix replication/errinj potential flakiness
The test has a function wait_repl() which tries to wait until the data
is replicated to the other node for a tiny timeout (0.2 seconds)

This didn't cause issues in the past, but it became an issue after
applier in thread introduction.

Let's use the default test_run:wait_cond() approach instead of a custom
one. This way we increase waiting timeout (making the test pass with a
higher chance).

Follow-up #6329